requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.8.3
aiosignal==1.2.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.0.1
attrs==21.4.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
boto3==1.24.45
botocore==1.27.45
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.4
cycler==0.11.0
dask==2023.1.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distlib==0.3.6
distro==1.7.0
docker-pycreds==0.4.0
dpath==2.0.6
dulwich==0.20.45
dvc==2.18.1
dvc-data==0.1.17
dvc-http==0.0.2
dvc-objects==0.1.11
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.10.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.9.0
flake8==6.0.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.7.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
h5py==3.7.0
huggingface-hub==0.11.1
hydra-core==1.2.0
identify==2.5.18
idna==3.3
imageio==2.31.1
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
lazy_loader==0.3
lime==0.2.0.1
llvmlite==0.40.1
locket==1.0.0
mariadb==1.0.11
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.2
nanotime==0.5.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.7
nodeenv==1.7.0
notebook==6.4.11
numba==0.57.1
numpy==1.22.3
omegaconf==2.2.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==3.0.0
pre-commit==3.0.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.1
pydot==1.4.2
pyflakes==3.0.1
pygit2==1.10.0
Pygments==2.12.0
pygtrie==2.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-transformers==1.2.0
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
rapidfuzz==2.13.3
regex==2022.7.25
requests==2.28.1
rich==12.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sacremoses==0.0.53
scikit-image==0.21.0
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.25
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.9.2
setproctitle==1.3.1
shap==0.42.1
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
slicer==0.0.7
smmap==5.0.0
soupsieve==2.3.2.post1
sshfs==2022.6.0
stack-data==0.2.0
swifter==1.3.4
sympy==1.12
tabulate==0.8.10
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2023.7.18
tinycss2==1.1.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.12.0
torchvision==0.15.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.25.1
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.11
vine==5.0.0
virtualenv==20.19.0
voluptuous==0.13.1
wandb==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.0
